refactor(nodes): Phase 1 unify lgnodes under nodes; rename to *.node.ts; delete obsolete file (refs #248, #263)
