+++ apps/server/__tests__/abort.propagation.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { LoggerService } from '../src/services/logger.service';
-import { ToolsNode } from '../src/lgnodes/tools.lgnode';
+import { ToolsNode } from '../src/nodes/tools.node';
 import { tool } from '@langchain/core/tools';
 import { AIMessage } from '@langchain/core/messages';
 import { z } from 'zod';
