+++ apps/server/__tests__/tools.node.terminate.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from 'vitest';
 import { AIMessage, BaseMessage } from '@langchain/core/messages';
-import { ToolsNode } from '../src/lgnodes/tools.lgnode';
+import { ToolsNode } from '../src/nodes/tools.node';
 import { BaseTool } from '../src/tools/base.tool';
 import { tool, DynamicStructuredTool } from '@langchain/core/tools';
 import { TerminateResponse } from '../src/tools/terminateResponse';
