+++ apps/server/src/nodes/base.node.ts
